uPy: fix blinking of R/G/B status LEDs due to context switching
Previously the code in main.py turned off all LEDs and then turned on
the needed one based on sensor readings. However, if execution context
is switched to background thread after LEDs have been turned off, but
before the new one was turned on, and the operation in background
thread was time-consuming (WiFi re-association, publishing to mqtt)
visible blinking of the LED could be observed.